refactor(scalars): move dropdown to new UI folder
