+++ packages/design-system/package.json
@@ -19,6 +19,10 @@
       "types": "./dist/src/scalars/index.d.ts",
       "import": "./dist/src/scalars/index.js"
     },
+    "./ui": {
+      "types": "./dist/src/ui/index.d.ts",
+      "import": "./dist/src/ui/index.js"
+    },
     "./style.css": "./dist/style.css"
   },
   "imports": {
