+++ packages/design-system/src/scalars/components/amount-field/use-amount-field.tsx
@@ -1,13 +1,13 @@
 import { useEffect, useMemo, useState } from "react";
-import { Currency } from "../currency-code-field/types.js";
+import { type Currency } from "../currency-code-field/types.js";
 import { isValidNumber } from "../number-field/number-field-validations.js";
 import {
-  Amount,
-  AmountCrypto,
-  AmountCurrency,
-  AmountFiat,
-  AmountFieldPropsGeneric,
-  AmountValue,
+  type Amount,
+  type AmountCrypto,
+  type AmountCurrency,
+  type AmountFiat,
+  type AmountFieldPropsGeneric,
+  type AmountValue,
 } from "./types.js";
 import {
   createAmountValue,
