+++ packages/design-system/src/scalars/components/fragments/index.ts
@@ -11,4 +11,3 @@ export * from "./text-field/index.js";
 export * from "./textarea-field/index.js";
 export * from "./toggle-field/index.js";
 export * from "./tooltip/index.js";
-
