+++ packages/design-system/src/scalars/components/index.ts
@@ -5,7 +5,6 @@ export * from "./country-code-field/index.js";
 export * from "./currency-code-field/index.js";
 export * from "./date-picker-field/index.js";
 export * from "./date-time-field/index.js";
-export * from "./dropdown/index.js";
 export * from "./enum-field/index.js";
 export * from "./form/index.js";
 export * from "./fragments/index.js";
@@ -19,3 +18,36 @@ export * from "./time-picker-field/index.js";
 export * from "./types.js";
 export * from "./url-field/index.js";
 
+// deprecated components
+export {
+  /**
+   * @deprecated Use the `Dropdown` component from the `@powerhousedao/design-system/ui` package instead.
+   * This component is deprecated and will be removed in a future version.
+   */
+  Dropdown,
+  /**
+   * @deprecated Use the `DropdownContent` component from the `@powerhousedao/design-system/ui` package instead.
+   * This component is deprecated and will be removed in a future version.
+   */
+  DropdownContent,
+  /**
+   * @deprecated Use the `DropdownItem` component from the `@powerhousedao/design-system/ui` package instead.
+   * This component is deprecated and will be removed in a future version.
+   */
+  DropdownItem,
+  /**
+   * @deprecated Use the `DropdownLabel` component from the `@powerhousedao/design-system/ui` package instead.
+   * This component is deprecated and will be removed in a future version.
+   */
+  DropdownLabel,
+  /**
+   * @deprecated Use the `DropdownShortcut` component from the `@powerhousedao/design-system/ui` package instead.
+   * This component is deprecated and will be removed in a future version.
+   */
+  DropdownShortcut,
+  /**
+   * @deprecated Use the `DropdownTrigger` component from the `@powerhousedao/design-system/ui` package instead.
+   * This component is deprecated and will be removed in a future version.
+   */
+  DropdownTrigger,
+} from "../../ui/components/dropdown/index.js";
