+++ packages/design-system/src/scalars/components/sidebar/subcomponents/sidebar-search/index.ts
@@ -1,4 +1,3 @@
 export { SearchNavigationArrow } from "./search-navigation-arrow.js";
 export { SearchResultCounter } from "./search-result-counter.js";
 export { SidebarSearch } from "./sidebar-search.js";
-
