+++ packages/design-system/src/ui/components/dropdown/subcomponents/dropdown-item.tsx
@@ -1,6 +1,6 @@
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu";
 import React from "react";
-import { cn } from "../../../lib/utils.js";
+import { cn } from "../../../../scalars/lib/utils.js";
 
 export const DropdownItem = React.forwardRef<
   React.ElementRef<typeof DropdownMenuPrimitive.Item>,
