+++ packages/design-system/src/ui/components/dropdown/subcomponents/dropdown-shortcut.tsx
@@ -1,4 +1,4 @@
-import { cn } from "../../../lib/utils.js";
+import { cn } from "../../../../scalars/lib/utils.js";
 
 export const DropdownShortcut = ({
   className,
