+++ packages/design-system/src/ui/components/index.ts
@@ -1,0 +1,11 @@
+// export all components accessible outside the package
+
+// dropdown
+export {
+  Dropdown,
+  DropdownContent,
+  DropdownItem,
+  DropdownLabel,
+  DropdownShortcut,
+  DropdownTrigger,
+} from "./dropdown/index.js";
